3B done, but buggy

+++ outLabs/outlab2/solution/Task3/viewWithColor.sh
@@ -1,30 +1,13 @@
 #!/bin/bash
 
-# source ./resources/defineColors.sh ########################################
+source ./resources/defineColors.sh 
 
-BLACK_FONT=`tput setaf 0`
-RED_FONT=`tput setaf 1`
-GREEN_FONT=`tput setaf 2`
-YELLOW_FONT=`tput setaf 3`
-BLUE_FONT=`tput setaf 4`
-MAGENTA_FONT=`tput setaf 5`
-CYAN_FONT=`tput setaf 6`
-WHITE_FONT=`tput setaf 7`
-
-BLACK_BACKGROUND=`tput setab 0`
-RED_BACKGROUND=`tput setab 1`
-GREEN_BACKGROUND=`tput setab 2`
-YELLOW_BACKGROUND=`tput setab 3`
-BLUE_BACKGROUND=`tput setab 4`
-MAGENTA_BACKGROUND=`tput setab 5`
-CYAN_BACKGROUND=`tput setab 6`
-WHITE_BACKGROUND=`tput setab 7`
-
-RESET_ALL=`tput sgr0`
-
-
-rm back.out
-rm font.out
+if [ -f "back.out" ]; then
+	rm back.out
+fi
+if [ -f "font.out" ]; then
+	rm font.out
+fi
 
 NRnow=$( awk '
 BEGIN {
@@ -69,18 +52,17 @@ END {
 }' $2 $1)
 
 
-
 for (( i = 1; i <= NRnow; i++ )); do
 
 	CURR_FONT=$(awk -v iter="$i" 'BEGIN{FS="\n"}{if(NR == iter) {print;}}' font.out)
 	CURR_BACk=$(awk -v iter="$i" 'BEGIN{FS="\n"}{if(NR == iter) {print;}}' back.out)
 
-	NewFont=${CURR_FONT}
-	NewBack=${CURR_BACk}
-	NewReset=${RESET_ALL}
+
+	NewFont=${!CURR_FONT}
+	NewBack=${!CURR_BACk}
 
 
-	awk -v toprint="$i" -v Font=$NewFont -v Back=$NewBack -v Reset=$NewReset '
+	awk -v toprint="$i" -v Font="$NewFont" -v Back="$NewBack" '
 	BEGIN {
 	} 
 	{
@@ -89,13 +71,14 @@ for (( i = 1; i <= NRnow; i++ )); do
 		} else if (NR == toprint) {
 			printf "" Font;
 			printf "" Back;
-			# printf("'$Font'");
-			# printf("'$Back'");
 			print;
+			printf("'$RESET_ALL'")
 		}
 	}
 	END {
-		printf Reset;
 	}
 	' $1
 done
+
+rm back.out
+rm font.out
